add is_surfaced_category_type to export_archives.insights_supervision_officer_outlier_status_archive schema (Recidiviz/recidiviz-data#40937)
## Description of the change

This is needed in order to get the new field from
https://github.com/Recidiviz/recidiviz-data/pull/40793 in the archive
views. I wasn't sure if an export with the new field needed to have
happened already in order to add this, but I verified by creating a
table with this schema manually in the bq UI against the prod data
(which wouldn't have this field in any exports) and it created just
fine.

This is part of bullet point 2 in
https://github.com/Recidiviz/recidiviz-data/issues/40790#issuecomment-2798196623.
Since the view schema won't have the `is_surfaced_category_type` field
until after this change is deployed, I think this is the last change
that I can safely merge before the release candidate is cut.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

https://github.com/Recidiviz/recidiviz-data/issues/40790

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4c592c415d9cac69dfb84d794568c5c29dfec2ac