[raw data] add reminder message to known value enum errors (Recidiviz/recidiviz-data#40425)
## Description of the change

this PR adds additional context to known value validation errors; for
raw file configs that are not used in ingest views --
```
No ingest views references [test_file_tag]. To resolve this error, you can: (1) update the list of known_values for [sad_col]
if you want to keep the addition of new values as import-blocking; (2) add an import-blocking exclusion if you want to
keep the existing documentation but know that it may quickly become stale; or (3) remove all known_values for [sad_col]
from the yaml config to remove its designation as an enum.
```

for one that are used in ingest views
```
The following ingest views reference [tagBasicData]:
	-basic
	-tagBasicData
 If this column is used in an ingest enum mapping, adding the new values will help ensure that the enum failure occurs at
raw data import time instead of ingest time. If it is not used, you can (1) add the new values to list of known_values for
[sad_col] if you want to keep the addition of new values as import-blocking; (2) add an import-blocking exclusion if you
want to keep the existing documentation but know that it may quickly become stale; or (3) remove all known_values for
[sad_col] from the yaml config to remove its designation as an enum.
```
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#40233

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a0ed66ae63e70168ef9b282c1495880962948432