[US_TN] Update almost-eligibility logic for SDS (Recidiviz/recidiviz-data#39929)
## Description of the change

Updated the almost-eligibility logic for TN SDS to add another pathway
for clients to be surfaced as almost-eligible. We'll now surface people
a short period of time in advance of eligibility (as long as they meet
the fines/fees criterion already or have a limited balance).

To keep the query complexity down for the SDS eligibility spans, I had
to group the two time-based criteria that we use in the
almost-eligibility logic. That combined criterion is written in a way
that should ensure that we correctly handle missing and null dates in
reasons blobs, which wouldn't necessarily be the case if we used an
`AndCriteriaGroup`.

I also set up `metadata_tab_name` in the opportunity record so that
we'll be able to subcategorize the almost-eligible clients in the tool,
making it easier for POs to identify who's approaching their eligibility
date vs. who's already met the accrued-time criteria but is just missing
the fines/fees criterion.

To load relevant views to sandbox:
```
python -m recidiviz.tools.load_views_to_sandbox \
--sandbox_dataset_prefix mhilderbran_us_tn_sds_almost_eligible --state_code_filter US_TN --prompt auto \
--load_up_to_addresses workflows_views.us_tn_suspension_of_direct_supervision_record
```

As a result of the changes in this PR, ~35 more people will become
almost-eligible for SDS:
```
WITH main AS (
  SELECT
    state_code,
    person_id,
    is_eligible AS is_eligible_main,
    is_almost_eligible AS is_almost_eligible_main,
  FROM `recidiviz-staging.task_eligibility_spans_us_tn.suspension_of_direct_supervision_materialized`
  WHERE CURRENT_DATE BETWEEN start_date AND COALESCE(DATE_SUB(end_date, INTERVAL 1 DAY), '9999-12-31')
),
sandbox AS (
  SELECT
    state_code,
    person_id,
    is_eligible AS is_eligible_sandbox,
    is_almost_eligible AS is_almost_eligible_sandbox,
  FROM `recidiviz-staging.mhilderbran_us_tn_sds_almost_eligible_task_eligibility_spans_us_tn.suspension_of_direct_supervision_materialized`
  WHERE CURRENT_DATE BETWEEN start_date AND COALESCE(DATE_SUB(end_date, INTERVAL 1 DAY), '9999-12-31')
)
SELECT
  main.is_eligible_main,
  sandbox.is_eligible_sandbox,
  main.is_almost_eligible_main,
  sandbox.is_almost_eligible_sandbox,
  COUNT(*),
FROM main
FULL OUTER JOIN sandbox
  USING (state_code, person_id)
WHERE (main.is_eligible_main!=sandbox.is_eligible_sandbox)
  OR (main.is_almost_eligible_main!=sandbox.is_almost_eligible_sandbox)
GROUP BY 1, 2, 3, 4
```

We can check that the correct tab names show up in `metadata_tab_name`
in the opportunity record for almost-eligible clients now:
```
SELECT *
FROM `recidiviz-staging.mhilderbran_us_tn_sds_almost_eligible_workflows_views.us_tn_suspension_of_direct_supervision_record_materialized`
WHERE is_almost_eligible
```

## Related issues

Relates to Recidiviz/recidiviz-data#38270

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1bba67ae653ee225c9fc7547d0e80595901d6bb9