[US_UT] add sync management script (Recidiviz/recidiviz-data#41248)
## Description of the change

ever since carlos added a LOT more tables to their sync (~1900), we no
longer really want to copy all non-bopp tables by default as this can
take quite a long time; instead, there is a table
`metadata.tables_to_sync` that maintains the list of tables we sync
every day. this means, however, that if we want to add or remove tables,
we need to do so manually. this script is designed to help this process,
letting us add or remove specific tables, as well as add newly created
files tags.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

n/a

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 07b37cc53aa63f24a36e91eaf0640be320d3aa2c