migrate insights_supervision_officer_outlier_status_archive definition into yaml_managed + delete from TF (Recidiviz/recidiviz-data#41316)
## Description of the change

As per
https://recidiviz.slack.com/archives/C028X32LRH7/p1745013379311539?thread_ts=1745013144.796639&cid=C028X32LRH7.

It seems like today's deploy succeeded without this change, but it
doesn't hurt to give it a try and make sure we don't have confusion for
next week's prod deploy. If it works well, then I can move the other
`export_archives` files over.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 96ca2188e82189813b7ccdb84e7aa89518b738dc