[US_TX] Update critically understaffed location upload with new assumptions (Recidiviz/recidiviz-data#41317)
## Description of the change

The agency/office location data did not join nicely between the data we
currently ingest and the excel files for manual upload. This PR updates
the schema and our script to account for this.

So now the person uploading this data will need to:
- Add the region and district codes for the relevant DPO
- Add the year and month values to the sheet
- Remove the region summary rows
- Export the file as a CSV to upload

I thought since we are doing manual work to get this in, we might as
well do all of it in one place rather than making odd SQL logic
somewhere.

Once this is in, the combination of `OFFC_REGION` and `OFFC_DISTRICT`
will be able to join to `Staff.Agency_of_Employment` (the location
external ID in Staff Location Periods).

I also added enforced `dtype` to the script, because pandas was silently
adding decimals to our integers before uploading :/

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| **Type: Non-breaking refactor** | change addresses some tech debt item
or prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#41017

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9a413e6eb69e03b0da5aa573b5e6f326e7f2e4ab