[Workflows] Output person_id for every opportunity record (Recidiviz/recidiviz-data#41286)
## Description of the change

As part of work to address Recidiviz/recidiviz-dashboardsRecidiviz/recidiviz-data#8065, which is
a bug with how we choose which external id to display to users, I want
to dramatically simplify the number of places in our opportunity record
and client record queries where we rely on person external id values. We
really should not be using these values in joins that link various
person-level and client-level data together in the `client_record` /
`resident_record` queries and instead should use the person_id wherever
possible.

This PR takes a first step towards that goal by exposing person_id as a
column in every opportunity record query. These person_id values will
get loaded to Firestore (and could be used in the future if we wanted
to), but will not actually be read by frontend code.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-dashboardsRecidiviz/recidiviz-data#8065

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: dd2989b801e099a0fc88176f069688e220706ed6