[US_AZ] Fix bugs in supervision periods view (Recidiviz/recidiviz-data#41291)
## Description of the change

This addresses two bugs in the supervision periods ingest view query:
1. Filter supervision level updates to those that have an UPDT_DTM value
of _greater than 2019-12-01_, instead of greater than 2019-11-30. The
November date was the date of the system migration on which all of the
closed periods' rows were updated with an UPDT_DTM of the migration
date. This was causing us to track "updates" to supervision levels that
happened years after the period itself ended. Filtering out all updates
that happened _on_ the migration date fixes this problem.
2. We are tracking the supervision level assigned upon intake and all
subsequent updates to that supervision level in separate CTEs.
Previously, we were pulling the initial supervision level in both CTEs,
but using different date fields to gauge when that assignment took
place. This was causing an issue that looked similar to the one above -
periods were held open far longer than was appropriate because of a
trailing supervision level update that was not legitimate. This PR
tweaks the supervision level-tracking logic to only track rows that
contain an update from the originally-assigned supervision level to
avoid this duplication.

[Successful sandbox
pipeline](https://console.cloud.google.com/dataflow/jobs/us-central1/2025-04-21_09_58_30-12159601890296516196?project=recidiviz-staging)

I ran an end-to-end sandbox and confirmed that this brings the failure
level of the overlapping_supervision_periods validation from 23,543 down
to 198.
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2ca64695f30d015d6e8b320ef0d1e265763eaa8b