[US_TN] Enforce CellBedAssignment start date before end date in IP view (Recidiviz/recidiviz-data#40574)
## Description of the change

Some CellBedAssignment records have an assignment end date on or before
the assignment start date. This change filters them out before
constructing incarceration periods, because the dates don't make sense
and we don't know what the correct ones are. There are also a lot of
custody changes and unit changes inputted into the Classification and
CellBedAssignment tables dated after the latest OffenderMovement date.
We had already accounted for excluding these movements after a discharge
movement but this PR also adds functionality to exclude custody/unit
changes after releases to supervision and to court. If there is any new
movement in OffenderMovement after discharge, release to supervision, or
to court, the latest unit and custody information will be pulled in.
This fixes 11 of the 13 people surfaced in Recidiviz/recidiviz-data#37674. The remaining two
have TN data errors that I have reached out to ask about, and will need
to be addressed in some other way.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#37674

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: lydiamasri2 <lydia@recidiviz.org>
GitOrigin-RevId: eda792c855b2f629321176dee5e9f7c58b3f46f6