[Workflows][Denials] Initial process denials script (Recidiviz/recidiviz-data#40783)
## Description of the change

MI has requested analysis on their free text denial reasons. @cawarren
did the initial research for how to accomplish this and wrote this
script which feeds our client denial reasons data (both structured
reasons and free text reasons) through an LLM, asking it to better
categorize and summarize the reasons. The intention is for the LLM to
assign the structured user denial reasons as well as free form text into
pre-determined reasons clusters. We know that users sometimes assign the
wrong denial reason, miss a denial reason, or have a legitimate denial
reason that they log into free form text. We want to account for all of
this and end up with a clean set of clustered denial reasons data to run
some analysis on.

This is the original script written by @cawarren with a few adjustments
made for mypy/linting. I will continue to iterate on this script,
allowing it to read/write from a GCS bucket, and tweak the prompt to
attempt to catch a few edge cases, additional cleanup. I just wanted to
commit the original as a baseline!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues
Part of Recidiviz/recidiviz-data#73592

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 20e486e29365765d53b2acfc507431eaa184506e