[US_TN] Undo backdating workaround for CR completion event (Recidiviz/recidiviz-data#41357)
## Description of the change

Removed (at least temporarily but perhaps permanently) the
counter-backdating logic from the CR completion event in TN. Because we
hadn't yet made changes to add a backdating workaround for CR
eligibility, this PR will bring the completion event back in sync with
the eligibility spans (such that neither has any backdating workaround).
It should be consequently easier to do an overall assessment of the
impact of implementing backarounds vs. not doing so, as everything on
`main` can be consistent for CR and all the counter-backdating logic can
be in a separate branch for testing.

To load the state-agnostic `TRANSFER_TO_LIMITED_SUPERVISION`
completion-event view (which will once again have TN transitions now) to
sandbox:
```
python -m recidiviz.tools.load_views_to_sandbox --project_id recidiviz-staging \
--sandbox_dataset_prefix mhilderbran_us_tn_cr_undo_backdating_workaround manual \
--view_ids_to_load task_eligibility_completion_events_general.transfer_to_limited_supervision \
--update_ancestors False --update_descendants False
```

Validation notes:
- We can use the following query to confirm that the total number of
completion events per person doesn't change due to this PR.
```
WITH main AS (
  SELECT
    state_code,
    person_id,
    COUNT(*) AS n_events_main,
  FROM `recidiviz-staging.task_eligibility_completion_events_us_tn.transfer_to_limited_supervision_materialized`
  GROUP BY 1, 2
),
sandbox AS (
  SELECT
    state_code,
    person_id,
    COUNT(*) AS n_events_sandbox,
  FROM `recidiviz-staging.mhilderbran_us_tn_cr_undo_backdating_workaround_task_eligibility_completion_events_general.transfer_to_limited_supervision_materialized`
  WHERE state_code='US_TN'
  GROUP BY 1, 2
)
SELECT *
FROM main
FULL OUTER JOIN sandbox
  USING (state_code, person_id)
WHERE (n_events_main!=sandbox.n_events_sandbox)
  OR (n_events_main IS NULL AND n_events_sandbox IS NOT NULL)
  OR (n_events_main IS NOT NULL AND n_events_sandbox IS NULL)
```
- We can use the following query to check how the event dates have
changed for each person. NB: this query just compares the event dates
for each person according to the order of the dates (comparing their
earliest date in the main version to their earliest date in the sandbox,
their 4th date in main to their 4th date in the sandbox, and so on),
which means that in this particular query, we can't definitively link a
particular main date to the specific date that replaced it in the
sandbox. We see date changes that look consistent with undoing the
counter-backdating logic.
```
WITH main AS (
  SELECT
    state_code,
    person_id,
    ROW_NUMBER() OVER (PARTITION BY state_code, person_id ORDER BY completion_event_date) AS r,
    completion_event_date AS completion_event_date_main,
  FROM `recidiviz-staging.task_eligibility_completion_events_us_tn.transfer_to_limited_supervision_materialized`
),
sandbox AS (
  SELECT
    state_code,
    person_id,
    ROW_NUMBER() OVER (PARTITION BY state_code, person_id ORDER BY completion_event_date) AS r,
    completion_event_date AS completion_event_date_sandbox,
  FROM `recidiviz-staging.mhilderbran_us_tn_cr_undo_backdating_workaround_task_eligibility_completion_events_general.transfer_to_limited_supervision_materialized`
  WHERE state_code='US_TN'
)
SELECT
  state_code,
  person_id,
  completion_event_date_main,
  completion_event_date_sandbox,
  DATE_DIFF(completion_event_date_sandbox, completion_event_date_main, DAY) AS days_diff,
FROM main
FULL OUTER JOIN sandbox
  USING (state_code, person_id, r)
ORDER BY days_diff
```

## Related issues

Related to Recidiviz/recidiviz-data#40144

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: af252a64bdc0d7f1f4069c31b663c5ee0c060d54