[US_AZ] Account for manual lock on controlling offenses (Recidiviz/recidiviz-data#41244)
## Description of the change
The "final offense" field in raw data in AZ can be manually changed, and
that change gets stored in a separate field (`FINAL_OFFENSE_ML`, where
ML stands for Manual Lock). I wasn't checking for a value in that field
originally, so I was pulling sentence group dates that were associated
with the wrong offense (the `FINAL_OFFENSE_ID`, when there was a
distinct `FINAL_OFFENSE_ID_ML`).

This PR makes it so that now the sentence length and sentence group
length views will check if there is a value in that manual lock field,
and use it instead of the standard field if so. This issue was causing
the
[inferred_group_aggregated_projected_dates_validation](https://admin-panel-prod.recidiviz.org/admin/validation_metadata/status/details/inferred_group_aggregated_projected_dates_validation?stateCode=US_AZ)
to fail dramatically, but this fix brought the failure level down to 0
in a sandbox.

- 41,886 sentence groups have a manual lock on the final_offense_id that
was not previously being accounted for.
- 9,690 of those manual lock values are distinct from the original final
offense value, which would result in a change in ingested sentence group
length

I believe there was also a typo in the validation view query that was
comparing the sentence `parole_eligibility_date` to the sentence group
`projected_parole_release_date`. I changed that here, but can easily
change back if it was intentional!

[Successful
pipeline](https://console.cloud.google.com/dataflow/jobs/us-central1/2025-04-18_09_14_25-12124063798955793596;graphView=0?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 03f55be2995c3d346d772d1788cfe7f061047953