Create TX Early Release from Supervision opportunity (Recidiviz/recidiviz-data#41268)
## Description of the change

Adding Early Release from Supervision as a new workflow in addition to
Annual Report Status. This opportunity also has new criteria in the new
policy TX is approving. Importantly, ERS does not fully release clients
from supervision, but eliminates any required contact for the rest of
their sentence.

As a part of this workflow, I created a new TaskCompletionEventType:
TRANSFER_TO_UNSUPERVISED_SUPERVISION . I think this is right because
this workflow is different from TRANSFER_TO_LIMITED_SUPERVISION (which
Annual Report Status uses) and it's not a Supervision Level Downgrade
(the client must be specifically transferred to "Unsupervised"
supervision level). But let me know if that's incorrect!

| Old Requirement | New Requirement |
|-----------------|-----------------|
| Have been under supervision for at least one half of the time that
remained on their sentence when released from prison | Have been under
supervision for at least one half of the time that remained on their
sentence when released to parole or mandatory supervision |
| Have been on supervision for two (2) years, plus one (1) year of
quarterly report status | Have been on supervision for two years or more
with a full year on low supervision |
| Have no current or prior felony convictions or deferred adjudication
for a non-qualifying offense (reference Parole Division Document Library
for the list of Non-qualifying Offenses) | Entirely removed |
| Have no warrant issued during the previous seven (7) years of the
current parole supervision period. This does not apply to a warrant
issued where subsequently an investigation or administrative review did
not sustain a violation | Have no warrant issued during the previous two
years of the current parole supervision period. This does not apply to a
warrant issued where subsequently an investigation or administrative
review did not sustain a violation |
| Have not committed any violation of rules or conditions of release as
indicated on their Parole Certificate, during the preceding two (2) year
period | Basically the same: Have not committed any violation of rules
or conditions of release as indicated on the release certificate, during
the preceding two year period |
| Have no current or prior felony convictions or deferred adjudication,
including juvenile convictions for the offense that includes the use of
a child in the commission of a crime, nor any offenses that were pled
down to a lesser degree but may have included an intent to commit sexual
assault, bodily harm, etc. | Same |

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f2a89c64e53ada1426c8396abd41c2cc0beca792